ai-image-detector/requirements.txt
# Required Dependencies
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pillow>=9.5.0

# Configuration Management
pyyaml>=6.0

# Optional Dependencies
# Diffusion-based noise extraction
diffusers>=0.21.0

# Spectral artifact visualization
pytorch-grad-cam>=1.4.0

# Evaluation metrics and clustering
scikit-learn>=1.3.0

# Visualization and logging
matplotlib>=3.7.0
tqdm>=4.65.0

# Development Dependencies
# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Property-based testing
hypothesis>=6.82.0
